# Conflicts:
#	MWSE.sln
#	MWSE/LuaDisableableEventManager.cpp
#	MWSE/LuaManager.cpp
#	MWSE/LuaObject.h
#	MWSE/MWSE.vcxproj.filters
#	MWSE/NIObject.cpp
#	MWSE/NIObjectLua.h
#	MWSE/NISourceTextureLua.cpp
#	MWSE/TES3ActorLua.h
#	MWSE/TES3LeveledListLua.cpp
#	MWSE/TES3MobileActorLua.h
#	MWSE/TES3Object.cpp
#	MWSE/TES3Object.h
#	MWSE/TES3ObjectLua.h
#	MWSE/TES3RaceLua.cpp
#	MWSE/TES3ReferenceLua.cpp
#	MWSE/TES3UtilLua.cpp
#	MWSE/TES3WeatherControllerLua.cpp

+++ MWSE/TES3Land.h
@@ -13,13 +13,13 @@ namespace TES3 {
 		int unknown_0x1C;
 		int unknown_0x20;
 		int unknown_0x24;
-		int unknown_0x28;
+		float unknown_0x28;
 		int unknown_0x2C[128];
 		NI::Pointer<NI::Object> unknown_0x22C;
 		int gridX; // 0x230
 		int gridY; // 0x234
-		int unknown_0x238;
-		int unknown_0x23C;
+		float unknown_0x238;
+		float unknown_0x23C;
 	};
 	static_assert(sizeof(Land) == 0x240, "TES3::Land failed size validation");
 }
